+++ tests/parts/test_presentation.py
@@ -146,6 +146,32 @@ class DescribePresentationPart(object):
         assert rId == "rId42"
         assert slide is slide_
 
+    def it_can_add_a_slide_cloned_from_another_presentation(
+        self, request, package_, slide_part_, slide_, relate_to_
+    ):
+        source_slide_ = instance_mock(request, Slide)
+        source_slide_part_ = instance_mock(request, SlidePart)
+        source_slide_.part = source_slide_part_
+        slide_layout_ = instance_mock(request, SlideLayout)
+        slide_layout_part_ = slide_layout_.part
+
+        partname = PackURI("/ppt/slides/slide9.xml")
+        property_mock(request, PresentationPart, "_next_slide_partname", return_value=partname)
+        SlidePart_ = class_mock(request, "pptx.parts.presentation.SlidePart")
+        SlidePart_.clone_from.return_value = slide_part_
+        slide_part_.slide = slide_
+        relate_to_.return_value = "rId42"
+        prs_part = PresentationPart(None, None, package_, None)
+
+        rId, slide = prs_part.add_slide_from_external(source_slide_, slide_layout_)
+
+        SlidePart_.clone_from.assert_called_once_with(
+            source_slide_part_, partname, package_, slide_layout_part_
+        )
+        prs_part.relate_to.assert_called_once_with(prs_part, slide_part_, RT.SLIDE)
+        assert rId == "rId42"
+        assert slide is slide_
+
     def it_finds_the_slide_id_of_a_slide_part(self, slide_part_, related_part_):
         prs_elm = element(
             "p:presentation/p:sldIdLst/(p:sldId{r:id=a,id=256},p:sldId{r:id="